logindefs: change getlogindefs_num() to return unsigned long
Where ever getlogindefs_num() is called return value is always
expected to be unsigned, such as sleep() input, gid_t or mode_t.

+++ login-utils/su.c
@@ -776,7 +776,7 @@ main (int argc, char **argv)
   if (!correct_password (pw))
     {
       log_su (pw, false);
-      sleep (getlogindefs_num ("FAIL_DELAY", 1));
+      sleep ((unsigned int)getlogindefs_num ("FAIL_DELAY", 1));
       error (EXIT_FAIL, 0, _("incorrect password"));
     }
   else
